Add support for setting port address.
This commit modifies the ESclient->init method to allow the
specification of a port to be used for CURL based REST connection
to the index engine.

The argument is provided in the form of a null-terminated string
buffer containing the ASCII representation of the Base 10 value
of the port to be used for the connection.