Implement consistent arguments for index engine utilities.
This commit modifies the following utilities:

quixote-inject
qidx-util

So that consistent arguments and environment variable support are
used.

Both utilities now use the -p command-line argument to specify an
alternate port to be used for the REST connection to the indexing
engine.

Both utilities now also use the -P command-line argument to
specify a password to be used for connecting to the engine.

Both utilities now also use common environment variables in order
to specify parameters to be used for the REST connections.

The following environment variables are implemented:

QUIXOTE_IDX_HOST
QUIXOTE_IDX_PORT
QUIXOTE_IDX_USER
QUIXOTE_IDX_PWD

Which should have self-explanatory roles.

The environment variables are used if values are not specified
through the use of command-line arguments.