Fix handling of default index access parameters.
A regression was noted with the qidx-util command where the -L
command-line argument to list the indexes in the endpoint was
failing with an error message indicating that the index was not
found.  If a request is made to list the indexes in an endpoint
it is obvious that there is no need to specify an index.

This was caused by the fact that the check for whether or not an
index was specified was conducted irregardless of the mode that
the utility was running in.  Gate the condition that checks
whether an index has been specified on whether or not the utility
was running in 'list' mode.  There is no need to check for the
specification of an index if the indexes are being listed.

The string expression of the port number to connect to was being
unconditionally set to the value of the QUIXOTE_IDX_PORT
environment variable.  This cause any value that was specified by
a command-line argument to be overridden.  Only set the port
value specified by the environment variable if a port
specification has not been declared as a command-line option.